chore: extract namespaced padding share
